modify

+++ cadastro_usuario.php
@@ -2,10 +2,14 @@
 
     include("conexao.php");
 
-    $pasta_destino = 'fotos/';
-    $extensao = strtolower(substr($_FILES['foto']['name'], -4));
-    $nome_foto = $pasta_destino . date("Ymd-His") . $extensao;
-    move_uploaded_file($_FILES['foto']['tmp_name'], $nome_foto);
+        $nome_foto = "";
+    if(file_exists($_FILES['foto']['tmp_name'])){
+        $pasta_destino = 'fotos/';
+        $extensao = strtolower(substr($_FILES['foto']['name'], -4));
+        $nome_foto = $pasta_destino . date("Ymd-His") . $extensao;
+        move_uploaded_file($_FILES['foto']['tmp_name'], $nome_foto);
+    }
+
